program: Always distribute full amount
Due to truncation during division, it's often possible to not distribute
the full `amount` provided to the instruction. To get around that, we
allow the treasury to get all of the lamports truncated during reward
calculation. That seemed like the fairest option, since otherwise
someone could call the distribution with just 1 lamport to screw over
the holders or the stakers.